[Init] (BE) : ERD 변경으로 엔티티 초기화
여기서부터 작업 바랍니다.

+++ BE/mvp/src/main/java/com/mvp/manager/repository/ManagerRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface ManagerRepository extends JpaRepository<Manager, Integer> {
+public interface ManagerRepository extends JpaRepository<Manager, Long> {
     Optional<Manager> findByEmail(String email);
 }
